requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.2.3
astroid==2.4.2
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.5.3
beautifulsoup4==4.9.1
botocore==1.17.15
Brotli==1.0.7
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
config==0.5.0.post0
Counter==1.0.0
dash
dash-core-components
dash-html-components
dash-bootstrap-components
dash-renderer
dash-table
DateTime==4.3
decouple==0.0.7
distlib==0.3.1
docutils==0.15.2
Faker==4.1.1
filelock==3.0.12
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
idna==2.9
importlib-metadata==1.7.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
nltk==3.5
numpy==1.19.0
packaging==20.4
panda==0.3.1
pandas==1.0.5
percy==2.0.2
pipenv==2020.6.2
plotly==4.8.2
pluggy==0.13.1
praw==7.1.0
prawcore==1.4.0
psycopg2-binary
psycopg2-binary==2.9.5
public==2019.4.13
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
pylint==2.5.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==5.4.3
pytest-mock==3.1.1
pytest-sugar==0.9.3
python-dateutil==2.6.1
python-memcached==1.59
pytz==2020.1
query-string==2019.4.13
redis==3.5.3
regex==2020.6.8
requests==2.24.0
retrying==1.3.3
rq==1.4.3
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
termcolor==1.1.0
text-unidecode==1.3
textblob==0.15.3
toml==0.10.1
tqdm==4.46.1
typed-ast==1.4.1
Unidecode==1.1.1
update-checker==0.17
urllib3==1.25.9
vaderSentiment==3.3.2
virtualenv==20.0.25
virtualenv-clone==0.5.4
waitress==1.4.4
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.1.0
zope.interface==5.1.0
gunicorn== 19.9.0
celery==4.4.6
invoke==1.4.1
sqlalchemy==1.3.18
